requirements.txt
flask
selenium
gspread
google-auth
sendgrid
pandas
python-dotenv
webdriver-manager